LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-channelmergernode-interface/audiochannelmerger-disconnect-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "silent-disconnect" 
PASS Audit report 
PASS > [silent-disconnect]  
FAIL X Channel #0: Expected 1 for all values but found 10914 unexpected values: 
	Index	Actual
	[11136]	0
	[11137]	0
	[11138]	0
	[11139]	0
	...and 10910 more errors. assert_true: expected true got false
PASS   Channel #1 contains all the expected values in the correct order: [1,0]. 
PASS   The index of first zero in the channel #1 is equal to 11136. 
FAIL < [silent-disconnect] 1 out of 3 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 1 tasks were failed. assert_true: expected true got false
